+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Better error handling in `Update-Profile`
+- Imports codaamok module before updating profile
 
 ## [3.0.5] - 2021-11-13
 ### Fixed